+++ src/GameOverPvPMenu.h
@@ -24,7 +24,7 @@ private:
 	bool is_newrecord = false;
 
 public:
-	GameOverPvPMenu(sf::RenderWindow& win, sf::Font& font, Audio& audio, sf::Sprite& backgroundSprite, int max_size = 2, std::initializer_list<std::string> init_list = { "PLAY AGAIN", "BACK TO MENU" }, float y_pos = 4.0f);
+	GameOverPvPMenu(sf::RenderWindow& win, sf::Font& font, Audio& audio, sf::Sprite& backgroundSprite, uint8_t max_size = 2, std::initializer_list<std::string> init_list = { "PLAY AGAIN", "BACK TO MENU" }, float y_pos = 4.0f);
 
 	void draw() override; // переопределяем метод для отображения результата игры
 	GameState getNextState() override;
