+++ src/Slider.h
@@ -8,10 +8,10 @@ private:
 	sf::RenderWindow& window;
 	sf::RectangleShape scrollBar; // ползунок
 	sf::RectangleShape background; // фон позунка
-	int value = 0; // текущее значение
-	int max_value = 100; // максимальное значение
-	int step = 5; // шаг на который изменяется значение
-	int length = 0; // текущая длина ползунка
+	uint8_t value = 0; // текущее значение
+	uint8_t max_value = 100; // максимальное значение
+	uint8_t step = 5; // шаг на который изменяется значение
+	uint16_t length = 0; // текущая длина ползунка
 	sf::Vector2f size; // размеры ползунка
 
 	void convert(); // по value посчитать длину ползунка: [0, 100] -> [0, size.x]
@@ -20,13 +20,13 @@ private:
 public:
 	Slider(sf::RenderWindow& win);
 
-	void setValue(int value);
-	void setMaxValue(int value);
-	void setStep(int step);
-	int getValue() { return value; }
+	void setValue(uint8_t value);
+	void setMaxValue(uint8_t value);
+	void setStep(uint8_t step);
+	uint8_t getValue() { return value; }
 	void increase();
 	void decrease();
-	void setPosition(int x, int y);
+	void setPosition(float x, float y);
 	void setSize(sf::Vector2f size);
 	void drawBar();
 };
